Create a palette using colors extracted from an image

create_pal creates a colorblind-friendly palette using colors extracted from an image.

create_pal(
  image,
  n,
  resize = 0.1,
  method = "kmeans",
  colorblind = T,
  sort = "none",
  show.pal = T,
  title,
  ...
)

Arguments

image
a path to the raster image (JPG, JPEG, PNG, TIFF) from which the colors are to be extracted.

n
a positive integer. The number of colors in the palette.

resize
a number between 0 and 1. This indicates the fraction by which the width/height (pixels) of the original image is resized while keeping the aspect ratio. Default to 0.1.

method
the clustering method for classifying pixels into groups based on the RGB values. Options are "kmeans" and "Gaussian_mix". Default to "kmeans".

colorblind
logical. Whether to render the palette colorblind-friendly (also see the ‘Details’ section). Default to TRUE.

sort
a character indicating how the colors should be sorted. Options are "none", "hue", "saturation", and "value". Default to "none"(unsorted).

show.pal
logical. Whether to display the palette or not. Default to TRUE.

title
a character string giving the title of the displayed palette.

...
additional arguments passed to ggplot2::theme.

Value

A vector of hexadecimal color codes.

Details

Two clustering methods are available. For method = "kmeans", pixels are partitioned into clusters using kmeans, and the RGB values of the cluster centers are converted into the corresponding hexadecimal color codes. For method = "Gaussian_mix", pixel components are identified via multivariate Gaussian mixture modeling using ClusterR::GMM, and the RGB values of the component centroids are converted into the corresponding hexadecimal color codes.

If "colorblind = T", the original colors are replaced with colorblind-friendly colors using colorBlindness::replacePlotColor.

The colors in the palette can be sorted based on their HSV values. If sort = "hue", the colors are sorted by hue in an ascending order. If sort = "saturation", the colors are sorted by saturation in a descending order. If sort = "value", the colors are sorted by value in a descending order.

Examples

if (FALSE) {
library(PalCreatoR)
image_path <- system.file("Mountain.JPG", package = "PalCreatoR")

My_pal <- create_pal(image = image_path,
                     n = 5,
                     resize = 0.1,
                     method = "kmeans",
                     colorblind = T,
                     sort = "value",
                     show.pal = T,
                     title = "My Palette")

print(My_pal)}